Fix cal3d build on Mac. Not fixing buildbot: The state has already been obliterated.
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@77818 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ src/cal3d/coremesh.cpp
@@ -17,7 +17,7 @@
 #include "cal3d/coresubmesh.h"
 #include "cal3d/coresubmorphtarget.h"
 
-unsigned int CalCoreMesh::size() const {
+size_t CalCoreMesh::size() const {
     unsigned int r = sizeof(CalCoreMesh);
     std::vector<CalCoreSubmesh *>::const_iterator iter1;
     for (iter1 = m_vectorCoreSubmesh.begin(); iter1 != m_vectorCoreSubmesh.end(); ++iter1) {
